Move filters out of mod.rs
